Split core/memory_ops.py into write/delete/recall modules
Pure refactor, no behavior changes. Extracted via exact index-based
line-slicing (both moved bodies verified byte-identical to the
original before any import cleanup) -- same methodology as the
server.py/server_stdio.py module splits.

- core/memory_ops.py keeps the write surface (_update_memory,
  _store_memory, _replace_memory): 899 -> 344 lines.
- core/memory_delete.py (new): the delete surface (_delete_memory,
  _delete_impact, _remove_deleted_sources_from_summary,
  _restore_sources_from_deleted_summary, _delete_memories): 186 lines.
- core/memory_recall.py (new): the recall surface (_recall_exact,
  _recall_similar, _recall_text_search): 384 lines.
- core/memory.py: split its single memory_ops import block into 3
  statements, one per module.
- tests/test_hybrid_search.py, test_temporal_weighting.py,
  test_memory_database.py: repointed monkeypatch targets from
  memory_ops_module to memory_recall_module for the recall-surface
  names that moved (TEMPORAL_WEIGHT, RECALL_SCAN_LIMIT,
  _fetch_fts_candidate_ids, etc.) -- confirmed the old module no
  longer has these attributes, so a missed repoint would raise
  AttributeError rather than silently pass.
- endpoints/concepts.py: fixed a stale docstring reference to
  memory_ops.py for a pattern that moved to memory_delete.py.

Full suite: 607 passed, 28 skipped, 0 failed. ruff clean.
