+++ marm-mcp-server/tests/test_hybrid_search.py
@@ -190,10 +190,10 @@ async def test_recall_text_search_falls_back_to_like_when_fts5_raises(
         helper_called["value"] = True
         raise sqlite3.OperationalError("no such table: memories_fts")
 
-    from marm_mcp_server.core import memory_ops as memory_ops_module
+    from marm_mcp_server.core import memory_recall as memory_recall_module
 
     monkeypatch.setattr(
-        memory_ops_module, "_fetch_and_score_fts_rows", raise_operational_error
+        memory_recall_module, "_fetch_and_score_fts_rows", raise_operational_error
     )
 
     results = await memory.recall_text_search(
@@ -338,7 +338,7 @@ async def test_recall_similar_falls_back_to_semantic_scan_when_fts_returns_no_ca
     monkeypatch, tmp_path
 ):
     """Semantic fallback must return results when FTS finds no candidates."""
-    from marm_mcp_server.core import memory_ops as memory_ops_module
+    from marm_mcp_server.core import memory_recall as memory_recall_module
 
     memory = MARMMemory(str(tmp_path / "memory.db"))
     memory._encoder_failed = True
@@ -352,7 +352,7 @@ async def test_recall_similar_falls_back_to_semantic_scan_when_fts_returns_no_ca
             conn, "fallback-test", "semantic fallback content", vec
         )
 
-    monkeypatch.setattr(memory_ops_module, "_fetch_fts_candidate_ids", lambda *_: [])
+    monkeypatch.setattr(memory_recall_module, "_fetch_fts_candidate_ids", lambda *_: [])
 
     results = await memory.recall_similar(
         "fallback content", session="fallback-test", limit=5, query_vec=vec.copy()
@@ -366,7 +366,7 @@ async def test_recall_similar_falls_back_when_fts_candidates_have_no_embeddings(
     monkeypatch, tmp_path
 ):
     """When FTS candidates exist but none are scoreable, semantic fallback must run."""
-    from marm_mcp_server.core import memory_ops as memory_ops_module
+    from marm_mcp_server.core import memory_recall as memory_recall_module
 
     memory = MARMMemory(str(tmp_path / "memory.db"))
     memory._encoder_failed = True
@@ -382,7 +382,7 @@ async def test_recall_similar_falls_back_when_fts_candidates_have_no_embeddings(
 
     # Return a non-existent ID so ID-bounded fetch finds nothing scoreable
     monkeypatch.setattr(
-        memory_ops_module, "_fetch_fts_candidate_ids", lambda *_: ["non-existent-id"]
+        memory_recall_module, "_fetch_fts_candidate_ids", lambda *_: ["non-existent-id"]
     )
 
     results = await memory.recall_similar(
@@ -398,7 +398,7 @@ async def test_recall_similar_falls_back_when_fts_candidates_are_all_wrong_dimen
 ):
     """When FTS returns a real ID whose embedding has the wrong dimension, all candidates
     are dimension-skipped, similarities is empty, and semantic fallback must run."""
-    from marm_mcp_server.core import memory_ops as memory_ops_module
+    from marm_mcp_server.core import memory_recall as memory_recall_module
 
     mem = MARMMemory(str(tmp_path / "memory.db"))
     mem._encoder_failed = True
@@ -423,7 +423,7 @@ async def test_recall_similar_falls_back_when_fts_candidates_are_all_wrong_dimen
 
     # Force FTS to return only the wrong-dimension ID
     monkeypatch.setattr(
-        memory_ops_module, "_fetch_fts_candidate_ids", lambda *_: [wrong_id]
+        memory_recall_module, "_fetch_fts_candidate_ids", lambda *_: [wrong_id]
     )
 
     query_vec = correct_vec.copy()
@@ -441,7 +441,7 @@ async def test_recall_similar_falls_back_when_fts_candidates_are_all_wrong_dimen
 @pytest.mark.asyncio
 async def test_recall_similar_respects_fts_candidate_limit(monkeypatch, tmp_path):
     """Scorer receives no more than max(limit, FTS_CANDIDATE_LIMIT) IDs."""
-    from marm_mcp_server.core import memory_ops as memory_ops_module
+    from marm_mcp_server.core import memory_recall as memory_recall_module
 
     cap = 3
     recall_limit = 5
@@ -457,14 +457,16 @@ async def test_recall_similar_respects_fts_candidate_limit(monkeypatch, tmp_path
             _insert_with_embedding(conn, "cap-test", f"canary memory entry {i}", vec)
 
     scored_ids: list[str] = []
-    original_scorer = memory_ops_module._fetch_and_score_by_ids
+    original_scorer = memory_recall_module._fetch_and_score_by_ids
 
     def capturing_scorer(db_path, memory_ids, query_embedding):
         scored_ids.extend(memory_ids)
         return original_scorer(db_path, memory_ids, query_embedding)
 
-    monkeypatch.setattr(memory_ops_module, "FTS_CANDIDATE_LIMIT", cap)
-    monkeypatch.setattr(memory_ops_module, "_fetch_and_score_by_ids", capturing_scorer)
+    monkeypatch.setattr(memory_recall_module, "FTS_CANDIDATE_LIMIT", cap)
+    monkeypatch.setattr(
+        memory_recall_module, "_fetch_and_score_by_ids", capturing_scorer
+    )
 
     await mem.recall_similar(
         "canary", session="cap-test", limit=recall_limit, query_vec=vec.copy()
@@ -481,7 +483,7 @@ async def test_recall_similar_fts_filter_failure_falls_back_to_semantic_scan(
     monkeypatch, tmp_path
 ):
     """An exception from _fetch_fts_candidate_ids must not crash recall."""
-    from marm_mcp_server.core import memory_ops as memory_ops_module
+    from marm_mcp_server.core import memory_recall as memory_recall_module
 
     memory = MARMMemory(str(tmp_path / "memory.db"))
     memory._encoder_failed = True
@@ -498,7 +500,9 @@ async def test_recall_similar_fts_filter_failure_falls_back_to_semantic_scan(
     def raise_fts_error(*args, **kwargs):
         raise sqlite3.OperationalError("no such table: memories_fts")
 
-    monkeypatch.setattr(memory_ops_module, "_fetch_fts_candidate_ids", raise_fts_error)
+    monkeypatch.setattr(
+        memory_recall_module, "_fetch_fts_candidate_ids", raise_fts_error
+    )
 
     results = await memory.recall_similar(
         "content", session="fts-fail", limit=5, query_vec=vec.copy()
@@ -549,12 +553,12 @@ async def test_recall_similar_debug_logs_filter_rerank_path(monkeypatch, tmp_pat
             conn, "debug-rerank", "docker deployment debug", vec
         )
 
-    from marm_mcp_server.core import memory_ops as memory_ops_module
+    from marm_mcp_server.core import memory_recall as memory_recall_module
 
     debug_calls: list[str] = []
-    monkeypatch.setattr(memory_ops_module, "_recall_debug", debug_calls.append)
+    monkeypatch.setattr(memory_recall_module, "_recall_debug", debug_calls.append)
     monkeypatch.setattr(
-        memory_ops_module, "_fetch_fts_candidate_ids", lambda *_: [embed_id]
+        memory_recall_module, "_fetch_fts_candidate_ids", lambda *_: [embed_id]
     )
 
     await mem.recall_similar(
@@ -581,11 +585,11 @@ async def test_recall_similar_debug_logs_semantic_fallback_path(monkeypatch, tmp
             conn, "debug-fallback", "semantic fallback debug content", vec
         )
 
-    from marm_mcp_server.core import memory_ops as memory_ops_module
+    from marm_mcp_server.core import memory_recall as memory_recall_module
 
     debug_calls: list[str] = []
-    monkeypatch.setattr(memory_ops_module, "_recall_debug", debug_calls.append)
-    monkeypatch.setattr(memory_ops_module, "_fetch_fts_candidate_ids", lambda *_: [])
+    monkeypatch.setattr(memory_recall_module, "_recall_debug", debug_calls.append)
+    monkeypatch.setattr(memory_recall_module, "_fetch_fts_candidate_ids", lambda *_: [])
 
     await mem.recall_similar(
         "fallback debug", session="debug-fallback", limit=5, query_vec=vec.copy()
